requirements_A.txt
numpy==2.2.2
pandas==2.2.2
PyMySQL==1.1.0
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
shiboken6==6.8.2.1
six==1.17.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
tzdata==2025.1
